chore: reorgs remote mode exports
