--- ui/pages/remote-mode/setup/index.scss
@@ -1,5 +1,0 @@
-@use "design-system";
-
-.dropdown-token {
-  width: 100px;
-}
