+++ ui/pages/remote-mode/setup/index.ts
@@ -1,0 +1,2 @@
+export { RemoteModeSetupSwaps } from './setup-swaps';
+export { RemoteModeSetupDailyAllowance } from './setup-daily-allowance';
