+++ ui/pages/remote-mode/setup/setup-daily-allowance/index.ts
@@ -1,0 +1,1 @@
+export { default as RemoteModeSetupDailyAllowance } from './remote-mode-setup-daily-allowance.component';
