+++ ui/pages/remote-mode/setup/setup-swaps/index.ts
@@ -1,0 +1,1 @@
+export { default as RemoteModeSetupSwaps } from './remote-mode-setup-swaps.component';
