+++ ui/pages/remote-mode/setup/setup-swaps/remote-mode-setup-swaps.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
-import configureStore from '../../../store/store';
-import RemoteModeSetupDailyAllowance from './remote-mode-setup-daily-allowance.component';
-import testData from '../../../../.storybook/test-data';
+import configureStore from '../../../../store/store';
+import RemoteModeSetupSwaps from './remote-mode-setup-swaps.component';
+import testData from '../../../../../.storybook/test-data';
 
 import { InternalAccount } from '@metamask/keyring-internal-api';
 
@@ -29,7 +29,7 @@ const mockAccounts: [InternalAccount] = [
 
 export default {
   title: 'Pages/Vault/RemoteMode/Setup',
-  component: RemoteModeSetupDailyAllowance,
+  component: RemoteModeSetupSwaps,
   decorators: [
     (story) => (
       <Provider store={store}>
@@ -39,6 +39,4 @@ export default {
   ],
 };
 
-export const Default = () => (
-  <RemoteModeSetupDailyAllowance accounts={mockAccounts} />
-);
+export const Default = () => <RemoteModeSetupSwaps accounts={mockAccounts} />;
